вынес логику создания коммента в popupcomponent
